+++ CHANGELOG.rst
@@ -8,9 +8,15 @@ v2.0.0
 
 Unreleased
 ------------------
+Added
+ - Added `tcod.context` module.  You now have more options for making libtcod
+   controlled contexts.
+
 Changed
  - `EventDispatch.dispatch` can now return the values returned by the `ev_*`
    methods.  The class is now generic to support type checking these values.
+ - Event mouse coordinates are now strictly int types.
+ - Submodules are now implicitly imported.
 
 11.11.4 - 2020-04-26
 --------------------